Add strapi headless cms
This commit adds strapi cms as an app running parallel to the client.
Both apps are connected via network on docker compose, so far.
The data from Strapi CMS are fetched on client server side, where it is loaded to the initialState and rendered before send upon resonse, which is benefitial for SEO puropses.